convert physics component
